python_basic_diploma/requirements.txt
python-dotenv==0.19.2
peewee==3.14.8
requests==2.26.0